requirements.txt
groq==1.1.2
google-generativeai==0.8.6
openai>=1.0.0
python-dotenv==1.2.2
nltk==3.9.4
